requirements.txt
bleach==1.5.0
boto3==1.4.5
botocore==1.5.92
certifi==2017.7.27.1
dj-database-url==0.4.2
Django==1.10.8
django-classy-tags==0.8.0
django-mptt==0.8.7
django-nyt==1.0b6
django-sekizai==0.10.0
django-storages==1.6.5
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
docutils==0.14
Faker==0.7.18
gunicorn==19.7.1
html5lib==0.9999999
jmespath==0.9.3
Markdown==2.6.10
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3
PyJWT==1.5.2
python-dateutil==2.6.1
pytz==2017.2
s3transfer==0.1.10
six==1.10.0
sorl-thumbnail==12.4.1
unicodecsv==0.14.1
